services/worker/requirements.txt
nats-py>=2.9.0
psycopg2-binary>=2.9.9
